Make trlGitInfo() call trlmessage() instead of just printf()
Resolves #198

Adds default macros for VERSION, BRANCH, & COMMIT to
include/hstcalversion.h that should have been part of #194

Fixes #233
Adds the missing dec ``char * sprintfGitInfo(char ** buffer);``
to hstcal/include/hstcalversion.h
